\name{MXM-internal}
\alias{InternalSES}
\alias{IdentifyEquivalence}
\alias{apply_ideq}
\alias{compare_p_values}
\alias{identifyTheEquivalent}
\alias{max_min_assoc}
\alias{min_assoc}
\alias{nchoosekm}
\alias{univariateScore}

\alias{proc_time-class}

\title{ Internal MXM Functions }
\description{
Internal functions of Package \pkg{MXM}
}
\section{Functions}{
\itemize{
\item InternalSES( ... )
\item IdentifyEquivalence( ... )
\item apply_ideq( ... )
\item compare_p_values( ... )
\item identifyTheEquivalent( ... )
\item max_min_assoc( ... )
\item min_assoc( ... )
\item nchoosekm( ... )
\item univariateScore( ... )
}
}
\details{
  These functions are only for internal usage of the MXM package - NOT to be called by the user.
  
  For faster computations in the internal SES functions, install the suggested package "\bold{gRbase}".
  %% or we have been too lazy to document
}
\keyword{Internal MXM Functions}